XLCellFormula separation 17/27 - Move CellFormula to outside scope because it must be accessible to load and save + rename to hungarian notation because of collision with DocumentFormat.OpenXml CellFormula class.
